Change MAX_STORIES to MAX_REPEATS
The number in that variable was used to determine how many times to go
through a repeating script. In the original code, the only repeating
script used was for stories, so calling it MAX_STORIES made sense; here,
we may repeat other kinds of scripts too, so we're making it more
general.