Remove unused code and attributes from the ScriptHandler
In particular, remove the database, question types, and the queue for
giving messages to the main interaction handler. We no longer keep
stories or performance in a database, we won't be using the same kind of
questions or question tracking, and we don't need to pause when we
timeout because there's no external game monitor.