Play back custom speech for the experiment, participant names
Allow the script files to include special tags for "experimenter_name"
and "particpant_name", which will get automatically filled in with the
name of the experimenter (passed in as a command line parameter on
startup) or the name of the participant (from the participant config
file). Either name is optional; if it is not provided, nothing will be
played.

While here, fix a few miscellaneous syntax errors.