Add script to check whether audio files in interaction scripts exist
The script uses the information in the main config.toml file to
determine which directories to check for scripts, and which directories
to check for the existence of audio and phoneme files. It will print a
list of all missing files to the screen (i.e., files referenced in the
script that don't exist as audio .wav or phoneme .txt files).