Make lookat node listen for robot sleep commands
- The lookat and affect node was still occasionally performing actions
when the robot was asleep; this is because it did not handle the case
where a sleep command was sent to the robot but a "robot sleeping" state
was not sent as an interaction state message (which may often be the
case if the user manually send sleep commands or if a script does not
happen to have that state command in it, which is easy to forget to
add). So now the lookat and affect node subscribes to tega commands so
it can see if the robot is being told to sleep or not.
- While here, fix a bug where the ROBOT_SLEEPING flag was never unset...
  now it will go false when the robot is no longer sleeping.