Don't do lookats, affect, or backchannel if robot is busy
Make the robot busy both when it is sleeping and when it is in a POSE
action. This will capture times when it is posing in a smile for a
picture as well as when it is sleeping, but won't suppress these actions
during other waiting questions (like during stories).