Add missing return to do_story_lookat
Without the early return in the if not RELATIONAL statement, the lookat
node was executing both story lookats for the relational and for the not
relational robots.... i.e. doing way too many! And this kept creating
more timer callbacks for lookats during the story, which led to way too
many threads being created, which eventually can crash roscore......