Add launching the interaction to the web interface
The web UI takes values and POSTs them to the server, which validates
them, then launches the interaction node in a subprocess.  The UI is
constantly updated with the node status.

Also add a restart arg "beginning" which has the effect of not providing
one.  It's easier to make it always have a value.